add sample HatsConditions contracts - #27

+++ src/HatsConditions/IHatsConditions.sol
@@ -3,5 +3,5 @@
 pragma solidity >=0.8.13;
 
 interface IHatsConditions {
-    function checkConditions(uint64 _hatId) external view returns (bool);
+    function checkConditions(uint256 _hatId) external view returns (bool);
 }
